Locales moved to services
