+++ lib/model/ui/plan/ui_plan_instance.dart
@@ -3,7 +3,7 @@ import 'package:bson/bson.dart';
 import 'package:fokus/model/db/plan/plan.dart';
 import 'package:fokus/model/db/plan/plan_instance.dart';
 import 'package:fokus/model/db/plan/plan_instance_state.dart';
-import 'package:fokus/utils/app_locales.dart';
+import 'package:fokus/services/app_locales.dart';
 
 import 'ui_plan_base.dart';
