+++ lib/utils/app_paths.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:fokus/model/currency_type.dart';
-import 'package:fokus/utils/app_locales.dart';
+import 'package:fokus/services/app_locales.dart';
 import 'package:fokus/utils/icon_sets.dart';
 
 String helpPagePath(BuildContext context, String helpPage) {
