+++ lib/widgets/app_navigation_bar.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:fokus/model/ui/app_page.dart';
 import 'package:fokus/model/ui/navigation_item.dart';
 import 'package:fokus/model/db/user/user.dart';
-import 'package:fokus/utils/app_locales.dart';
+import 'package:fokus/services/app_locales.dart';
 
 class AppNavigationBar extends StatefulWidget {
 	final int currentIndex;
