+++ lib/widgets/dialogs/help_dialog.dart
@@ -3,7 +3,7 @@ import 'package:flutter_markdown/flutter_markdown.dart';
 import 'package:flutter/services.dart' show rootBundle;
 
 import 'package:fokus/model/ui/button.dart';
-import 'package:fokus/utils/app_locales.dart';
+import 'package:fokus/services/app_locales.dart';
 import 'package:fokus/utils/app_paths.dart';
 
 class HelpDialog extends StatelessWidget {
