+++ lib/widgets/segment.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'package:fokus/utils/app_locales.dart';
+import 'package:fokus/services/app_locales.dart';
 import 'package:fokus/utils/dialog_utils.dart';
 import 'package:fokus/utils/theme_config.dart';
